+++ packages/did/src/DidDetails/LightDidDetails.ts
@@ -5,31 +5,52 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
+import type { DidDocument, DidUri } from '@kiltprotocol/types'
+
 import {
   base58Decode,
   base58Encode,
   decodeAddress,
 } from '@polkadot/util-crypto'
+import { cbor, SDKErrors, ss58Format } from '@kiltprotocol/utils'
 
 import type {
-  DidDocument,
-  DidServiceEndpoint,
-  DidUri,
-  LightDidSupportedVerificationKeyType,
   NewDidEncryptionKey,
-  NewLightDidVerificationKey,
-} from '@kiltprotocol/types'
-import { encryptionKeyTypes } from '@kiltprotocol/types'
+  NewDidVerificationKey,
+  NewService,
+  DidVerificationKeyType,
+} from './DidDetails.js'
 
-import { SDKErrors, ss58Format, cbor } from '@kiltprotocol/utils'
+import {
+  keypairToMultibaseKey,
+  didKeyToVerificationMethod,
+  getAddressFromVerificationMethod,
+  parse,
+} from '../Did.utils.js'
+import { fragmentIdToChain, validateNewService } from '../Did.chain.js'
+import {
+  addKeypairAsVerificationMethod,
+  encryptionKeyTypes,
+} from './DidDetails.js'
 
-import { getAddressByKey, parse } from '../Did.utils.js'
-import { resourceIdToChain, validateService } from '../Did.chain.js'
+/**
+ * Currently, a light DID does not support the use of an ECDSA key as its authentication key.
+ */
+export type LightDidSupportedVerificationKeyType = Extract<
+  DidVerificationKeyType,
+  'ed25519' | 'sr25519'
+>
+/**
+ * A new public key specified when creating a new light DID.
+ */
+export type NewLightDidVerificationKey = NewDidVerificationKey & {
+  type: LightDidSupportedVerificationKeyType
+}
+
+type LightDidEncoding = '00' | '01'
 
 const authenticationKeyId = '#authentication'
 const encryptionKeyId = '#encryption'
-
-type LightDidEncoding = '00' | '01'
 
 const verificationKeyTypeToLightDidEncoding: Record<
   LightDidSupportedVerificationKeyType,
@@ -52,12 +73,12 @@ const lightDidEncodingToVerificationKeyType: Record<
  */
 export type CreateDocumentInput = {
   /**
-   * The DID authentication key. This is mandatory and will be used as the first authentication key
+   * The key to be used as the DID authentication verification method. This is mandatory and will be used as the first authentication verification method
    * of the full DID upon migration.
    */
-  authentication: [NewLightDidVerificationKey]
+  authentication: [NewDidVerificationKey]
   /**
-   * The optional DID encryption key. If present, it will be used as the first key agreement key
+   * The optional encryption key to be used as the DID key agreement verification method. If present, it will be used as the first key agreement verification method
    * of the full DID upon migration.
    */
   keyAgreement?: [NewDidEncryptionKey]
@@ -65,22 +86,21 @@ export type CreateDocumentInput = {
    * The set of service endpoints associated with this DID. Each service endpoint ID must be unique.
    * The service ID must not contain the DID prefix when used to create a new DID.
    */
-  service?: DidServiceEndpoint[]
+  service?: NewService[]
 }
 
 function validateCreateDocumentInput({
   authentication,
   keyAgreement,
-  service: services,
+  service,
 }: CreateDocumentInput): void {
   // Check authentication key type
   const authenticationKeyTypeEncoding =
     verificationKeyTypeToLightDidEncoding[authentication[0].type]
 
-  if (!authenticationKeyTypeEncoding) {
+  if (authenticationKeyTypeEncoding === undefined) {
     throw new SDKErrors.UnsupportedKeyError(authentication[0].type)
   }
-
   if (
     keyAgreement?.[0].type &&
     !encryptionKeyTypes.includes(keyAgreement[0].type)
@@ -93,14 +113,14 @@ function validateCreateDocumentInput({
   // Checks that for all service IDs have regular strings as their ID and not a full DID.
   // Plus, we forbid a service ID to be `authentication` or `encryption` as that would create confusion
   // when upgrading to a full DID.
-  services?.forEach((service) => {
+  service?.forEach((s) => {
     // A service ID cannot have a reserved ID that is used for key IDs.
-    if (service.id === '#authentication' || service.id === '#encryption') {
+    if (s.id === '#authentication' || s.id === '#encryption') {
       throw new SDKErrors.DidError(
-        `Cannot specify a service ID with the name "${service.id}" as it is a reserved keyword`
+        `Cannot specify a service ID with the name "${s.id}" as it is a reserved keyword`
       )
     }
-    validateService(service)
+    validateNewService(s)
   })
 }
 
@@ -110,20 +130,20 @@ const SERVICES_MAP_KEY = 's'
 interface SerializableStructure {
   [KEY_AGREEMENT_MAP_KEY]?: NewDidEncryptionKey
   [SERVICES_MAP_KEY]?: Array<
-    Partial<Omit<DidServiceEndpoint, 'id'>> & {
+    Partial<Omit<NewService, 'id'>> & {
       id: string
     } & { types?: string[]; urls?: string[] } // This below was mistakenly not accounted for during the SDK refactor, meaning there are light DIDs that contain these keys in their service endpoints.
   >
 }
 
 /**
- * Serialize the optional encryption key and service endpoints of an off-chain DID using the CBOR serialization algorithm
+ * Serialize the optional key agreement key and service endpoints of a light DID using the CBOR serialization algorithm
  * and encoding the result in Base58 format with a multibase prefix.
  *
  * @param details The light DID details to encode.
- * @param details.keyAgreement The DID encryption key.
+ * @param details.keyAgreement The DID key agreement key.
  * @param details.service The DID service endpoints.
- * @returns The Base58-encoded and CBOR-serialized off-chain DID optional details.
+ * @returns The Base58-encoded and CBOR-serialized light DID optional details.
  */
 function serializeAdditionalLightDidDetails({
   keyAgreement,
@@ -136,7 +156,7 @@ function serializeAdditionalLightDidDetails({
   }
   if (service && service.length > 0) {
     objectToSerialize[SERVICES_MAP_KEY] = service.map(({ id, ...rest }) => ({
-      id: resourceIdToChain(id),
+      id: fragmentIdToChain(id),
       ...rest,
     }))
   }
@@ -188,8 +208,8 @@ function deserializeAdditionalLightDidDetails(
  * Private keys are assumed to already live in another storage, as it contains reference only to public keys.
  *
  * @param input The input.
- * @param input.authentication The array containing light DID authentication key.
- * @param input.keyAgreement The optional array containing light DID encryption key.
+ * @param input.authentication The array containing the public keys to be used as the light DID authentication verification method.
+ * @param input.keyAgreement The optional array containing the public keys to be used as the light DID key agreement verification methods.
  * @param input.service The optional light DID service endpoints.
  *
  * @returns The resulting [[DidDocument]].
@@ -211,32 +231,33 @@ export function createLightDidDocument({
   // Validity is checked in validateCreateDocumentInput
   const authenticationKeyTypeEncoding =
     verificationKeyTypeToLightDidEncoding[authentication[0].type]
-  const address = getAddressByKey(authentication[0])
+  const address = getAddressFromVerificationMethod({
+    publicKeyMultibase: keypairToMultibaseKey(authentication[0]),
+  })
 
   const encodedDetailsString = encodedDetails ? `:${encodedDetails}` : ''
   const uri =
     `did:kilt:light:${authenticationKeyTypeEncoding}${address}${encodedDetailsString}` as DidUri
 
   const did: DidDocument = {
-    uri,
-    authentication: [
-      {
-        id: authenticationKeyId, // Authentication key always has the #authentication ID.
-        type: authentication[0].type,
+    id: uri,
+    authentication: [authenticationKeyId],
+    verificationMethod: [
+      didKeyToVerificationMethod(uri, authenticationKeyId, {
+        keyType: authentication[0].type,
         publicKey: authentication[0].publicKey,
-      },
+      }),
     ],
     service,
   }
 
   if (keyAgreement !== undefined) {
-    did.keyAgreement = [
-      {
-        id: encryptionKeyId, // Encryption key always has the #encryption ID.
-        type: keyAgreement[0].type,
-        publicKey: keyAgreement[0].publicKey,
-      },
-    ]
+    const { publicKey, type } = keyAgreement[0]
+    addKeypairAsVerificationMethod(
+      did,
+      { id: encryptionKeyId, publicKey, type },
+      'keyAgreement'
+    )
   }
 
   return did
@@ -273,7 +294,7 @@ export function parseDocumentFromLightDid(
       `Cannot build a light DID from the provided URI "${uri}" because it does not refer to a light DID`
     )
   }
-  if (fragment && failIfFragmentPresent) {
+  if (fragment !== undefined && failIfFragmentPresent) {
     throw new SDKErrors.DidError(
       `Cannot build a light DID from the provided URI "${uri}" because it has a fragment`
     )
