--- packages/did/src/DidDetailsv2/index.ts
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2018-2023, BOTLabs GmbH.
- *
- * This source code is licensed under the BSD 4-Clause "Original" license
- * found in the LICENSE file in the root directory of this source tree.
- */
-
-// We don't export the `add*VerificationMethod` functions, they are meant to be used internally
-export {
-  BaseNewDidKey,
-  DidEncryptionKeyType,
-  DidKeyType,
-  DidVerificationKeyType,
-  NewDidEncryptionKey,
-  NewDidVerificationKey,
-  NewService,
-  NewVerificationMethod,
-} from './DidDetailsV2.js'
-export * from './LightDidDetailsV2.js'
-export * from './FullDidDetailsV2.js'
