--- packages/did/src/DidDocumentExporter/index.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2018-2023, BOTLabs GmbH.
- *
- * This source code is licensed under the BSD 4-Clause "Original" license
- * found in the LICENSE file in the root directory of this source tree.
- */
-
-export * from './DidDocumentExporter.js'
-export * from './DidContexts.js'
