+++ packages/did/src/DidLinks/AccountLinks.chain.ts
@@ -5,21 +5,10 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-import { decodeAddress, signatureVerify } from '@polkadot/util-crypto'
 import type { TypeDef } from '@polkadot/types/types'
 import type { KeypairType } from '@polkadot/util-crypto/types'
-import {
-  stringToU8a,
-  U8A_WRAP_ETHEREUM,
-  u8aConcatStrict,
-  u8aToHex,
-  u8aWrapBytes,
-  BN,
-} from '@polkadot/util'
-import { ApiPromise } from '@polkadot/api'
-
-import { SDKErrors } from '@kiltprotocol/utils'
-import { ConfigService } from '@kiltprotocol/config'
+import type { ApiPromise } from '@polkadot/api'
+import type { BN } from '@polkadot/util'
 import type {
   DidUri,
   HexString,
@@ -27,8 +16,18 @@ import type {
   KiltAddress,
 } from '@kiltprotocol/types'
 
-import { EncodedSignature } from '../Did.utils.js'
-import { toChain } from '../Did.chain.js'
+import { decodeAddress, signatureVerify } from '@polkadot/util-crypto'
+import {
+  stringToU8a,
+  U8A_WRAP_ETHEREUM,
+  u8aConcatStrict,
+  u8aToHex,
+  u8aWrapBytes,
+} from '@polkadot/util'
+import { SDKErrors } from '@kiltprotocol/utils'
+import { ConfigService } from '@kiltprotocol/config'
+
+import { toChain, EncodedSignature } from '../Did.chain.js'
 
 /**
  *  A chain-agnostic address, which can be encoded using any network prefix.
