+++ packages/did/src/DidResolver/DidResolver.ts
@@ -6,66 +6,75 @@
  */
 
 import type {
-  ConformingDidResolutionResult,
-  DidKey,
-  DidResolutionResult,
-  DidResourceUri,
+  DereferenceContentMetadata,
+  DereferenceContentStream,
+  DereferenceOptions,
+  DereferenceResult,
+  DidDocument,
+  DidResolver,
   DidUri,
-  KeyRelationship,
-  ResolvedDidKey,
-  ResolvedDidServiceEndpoint,
-  UriFragment,
+  DidUrl,
+  JsonLd,
+  RepresentationResolutionResult,
+  ResolutionDocumentMetadata,
+  ResolutionOptions,
+  ResolutionResult,
 } from '@kiltprotocol/types'
-import { SDKErrors } from '@kiltprotocol/utils'
-import { ConfigService } from '@kiltprotocol/config'
 
-import * as Did from '../index.js'
-import { toChain } from '../Did.chain.js'
+import { ConfigService } from '@kiltprotocol/config'
+import { cbor } from '@kiltprotocol/utils'
+
+import { KILT_DID_CONTEXT_URL, W3C_DID_CONTEXT_URL } from './DidContexts.js'
 import { linkedInfoFromChain } from '../Did.rpc.js'
-import { getFullDidUri, parse } from '../Did.utils.js'
-import { exportToDidDocument } from '../DidDocumentExporter/DidDocumentExporter.js'
+import { toChain } from '../Did.chain.js'
+import { getFullDidUri, parse, validateUri } from '../Did.utils.js'
+import { parseDocumentFromLightDid } from '../DidDetails/LightDidDetails.js'
+
+const DID_JSON = 'application/did+json'
+const DID_JSON_LD = 'application/did+ld+json'
+const DID_CBOR = 'application/did+cbor'
 
 /**
- * Resolve a DID URI to the DID document and its metadata.
- *
- * The URI can also identify a key or a service, but it will be ignored during resolution.
- *
- * @param did The subject's DID.
- * @returns The details associated with the DID subject.
+ * Supported types for DID resolution and dereferencing.
  */
-export async function resolve(
+export type SupportedContentType =
+  | typeof DID_JSON
+  | typeof DID_JSON_LD
+  | typeof DID_CBOR
+
+function isValidContentType(input: unknown): input is SupportedContentType {
+  return input === DID_JSON || input === DID_JSON_LD || input === DID_CBOR
+}
+
+type InternalResolutionResult = {
+  document?: DidDocument
+  documentMetadata: ResolutionDocumentMetadata
+}
+
+async function resolveInternal(
   did: DidUri
-): Promise<DidResolutionResult | null> {
+): Promise<InternalResolutionResult | null> {
   const { type } = parse(did)
   const api = ConfigService.get('api')
-  const queryFunction = api.call.did?.query ?? api.call.didApi.queryDid
-  const { section, version } = queryFunction?.meta ?? {}
-  if (version > 2)
-    throw new Error(
-      `This version of the KILT sdk supports runtime api '${section}' <=v2 , but the blockchain runtime implements ${version}. Please upgrade!`
-    )
-  const { document, web3Name } = await queryFunction(toChain(did))
-    .then(linkedInfoFromChain)
-    .catch(() => ({ document: undefined, web3Name: undefined }))
 
-  if (type === 'full' && document) {
+  const { document } = await api.call.did
+    .query(toChain(did))
+    .then(linkedInfoFromChain)
+    .catch(() => ({ document: undefined }))
+
+  if (type === 'full' && document !== undefined) {
     return {
       document,
-      metadata: {
-        deactivated: false,
-      },
-      ...(web3Name && { web3Name }),
+      documentMetadata: {},
     }
   }
 
-  // If the full DID has been deleted (or the light DID was upgraded and deleted),
-  // return the info in the resolution metadata.
   const isFullDidDeleted = (await api.query.did.didBlacklist(toChain(did)))
     .isSome
   if (isFullDidDeleted) {
     return {
       // No canonicalId and no details are returned as we consider this DID deactivated/deleted.
-      metadata: {
+      documentMetadata: {
         deactivated: true,
       },
     }
@@ -75,13 +84,15 @@ export async function resolve(
     return null
   }
 
-  const lightDocument = Did.parseDocumentFromLightDid(did, false)
+  const lightDocument = parseDocumentFromLightDid(did, false)
   // If a full DID with same subject is present, return the resolution metadata accordingly.
-  if (document) {
+  if (document !== undefined) {
     return {
-      metadata: {
+      documentMetadata: {
         canonicalId: getFullDidUri(did),
-        deactivated: false,
+      },
+      document: {
+        id: did,
       },
     }
   }
@@ -90,190 +101,224 @@ export async function resolve(
   // Metadata will simply contain `deactivated: false`.
   return {
     document: lightDocument,
-    metadata: {
-      deactivated: false,
-    },
+    documentMetadata: {},
   }
 }
 
 /**
  * Implementation of `resolve` compliant with W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
- * As opposed to `resolve`, which takes a more pragmatic approach, the `didDocument` property contains a fully compliant DID document abstract data model.
  * Additionally, this function returns an id-only DID document in the case where a DID has been deleted or upgraded.
  * If a DID is invalid or has not been registered, this is indicated by the `error` property on the `didResolutionMetadata`.
  *
  * @param did The DID to resolve.
- * @returns An object with the properties `didDocument` (a spec-conforming DID document or `undefined`), `didDocumentMetadata` (equivalent to `metadata` returned by [[resolve]]), as well as `didResolutionMetadata` (indicating an `error` if any).
+ * @param resolutionOptions The resolution options accepted by the `resolve` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
+ * @returns The resolution result for the `resolve` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
  */
-export async function resolveCompliant(
-  did: DidUri
-): Promise<ConformingDidResolutionResult> {
-  const result: ConformingDidResolutionResult = {
-    didDocumentMetadata: {},
-    didResolutionMetadata: {},
-  }
+export async function resolve(
+  did: DidUri,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  resolutionOptions: ResolutionOptions = {}
+): Promise<ResolutionResult> {
   try {
-    Did.validateUri(did, 'Did')
+    validateUri(did, 'Did')
   } catch (error) {
-    result.didResolutionMetadata.error = 'invalidDid'
-    if (error instanceof Error) {
-      result.didResolutionMetadata.errorMessage =
-        error.name + error.message ? `: ${error.message}` : ''
+    return {
+      didResolutionMetadata: {
+        error: 'invalidDid',
+      },
+      didDocumentMetadata: {},
     }
-    return result
   }
-  const resolutionResult = await resolve(did)
+
+  const resolutionResult = await resolveInternal(did)
   if (!resolutionResult) {
-    result.didResolutionMetadata.error = 'notFound'
-    result.didResolutionMetadata.errorMessage = `DID ${did} not found (on chain)`
-    return result
-  }
-  const { metadata, document, web3Name } = resolutionResult
-  result.didDocumentMetadata = metadata
-  result.didDocument = document
-    ? exportToDidDocument(document, 'application/json')
-    : { id: did }
-
-  if (web3Name) {
-    result.didDocument.alsoKnownAs = [`w3n:${web3Name}`]
+    return {
+      didResolutionMetadata: {
+        error: 'notFound',
+      },
+      didDocumentMetadata: {},
+    }
   }
 
-  return result
-}
-
-/**
- * Converts the DID key in the format returned by `resolveKey()`, useful for own implementations of `resolveKey`.
- *
- * @param key The DID key in the SDK format.
- * @param did The DID the key belongs to.
- * @returns The key in the resolveKey-format.
- */
-export function keyToResolvedKey(key: DidKey, did: DidUri): ResolvedDidKey {
-  const { id, publicKey, includedAt, type } = key
-  return {
-    controller: did,
-    id: `${did}${id}`,
-    publicKey,
-    type,
-    ...(includedAt && { includedAt }),
-  }
-}
-
-/**
- * Converts the DID key returned by the `resolveKey()` into the format used in the SDK.
- *
- * @param key The key in the resolveKey-format.
- * @returns The key in the SDK format.
- */
-export function resolvedKeyToKey(key: ResolvedDidKey): DidKey {
-  const { id, publicKey, includedAt, type } = key
-  return {
-    id: Did.parse(id).fragment as UriFragment,
-    publicKey,
-    type,
-    ...(includedAt && { includedAt }),
-  }
-}
-
-/**
- * Resolve a DID key URI to the key details.
- *
- * @param keyUri The DID key URI.
- * @param expectedVerificationMethod Optional key relationship the key has to belong to.
- * @returns The details associated with the key.
- */
-export async function resolveKey(
-  keyUri: DidResourceUri,
-  expectedVerificationMethod?: KeyRelationship
-): Promise<ResolvedDidKey> {
-  const { did, fragment: keyId } = parse(keyUri)
-
-  // A fragment (keyId) IS expected to resolve a key.
-  if (!keyId) {
-    throw new SDKErrors.DidError(
-      `Key URI "${keyUri}" is not a valid DID resource`
-    )
-  }
-
-  const resolved = await resolve(did)
-  if (!resolved) {
-    throw new SDKErrors.DidNotFoundError()
-  }
-
-  const {
-    document,
-    metadata: { canonicalId },
-  } = resolved
-
-  // If the light DID has been upgraded we consider the old key URI invalid, the full DID URI should be used instead.
-  if (canonicalId) {
-    throw new SDKErrors.DidResolveUpgradedDidError()
-  }
-  if (!document) {
-    throw new SDKErrors.DidDeactivatedError()
-  }
-
-  const key = Did.getKey(document, keyId)
-  if (!key) {
-    throw new SDKErrors.DidNotFoundError('Key not found in DID')
-  }
-
-  // Check whether the provided key ID is within the keys for a given verification relationship, if provided.
-  if (
-    expectedVerificationMethod &&
-    !document[expectedVerificationMethod]?.some(({ id }) => keyId === id)
-  ) {
-    throw new SDKErrors.DidError(
-      `No key "${keyUri}" for the verification method "${expectedVerificationMethod}"`
-    )
-  }
-
-  return keyToResolvedKey(key, did)
-}
-
-/**
- * Resolve a DID service URI to the service details.
- *
- * @param serviceUri The DID service URI.
- * @returns The details associated with the service endpoint.
- */
-export async function resolveService(
-  serviceUri: DidResourceUri
-): Promise<ResolvedDidServiceEndpoint> {
-  const { did, fragment: serviceId } = parse(serviceUri)
-
-  // A fragment (serviceId) IS expected to resolve a key.
-  if (!serviceId) {
-    throw new SDKErrors.DidError(
-      `Service URI "${serviceUri}" is not a valid DID resource`
-    )
-  }
-
-  const resolved = await resolve(did)
-  if (!resolved) {
-    throw new SDKErrors.DidNotFoundError()
-  }
-
-  const {
-    document,
-    metadata: { canonicalId },
-  } = resolved
-
-  // If the light DID has been upgraded we consider the old service URI invalid, the full DID URI should be used instead.
-  if (canonicalId) {
-    throw new SDKErrors.DidResolveUpgradedDidError()
-  }
-  if (!document) {
-    throw new SDKErrors.DidDeactivatedError()
-  }
-
-  const service = Did.getService(document, serviceId)
-  if (!service) {
-    throw new SDKErrors.DidNotFoundError('Service not found in DID')
-  }
+  const { documentMetadata: didDocumentMetadata, document: didDocument } =
+    resolutionResult
 
   return {
-    ...service,
-    id: `${did}${serviceId}`,
+    didResolutionMetadata: {},
+    didDocumentMetadata,
+    didDocument,
   }
+}
+
+/**
+ * Implementation of `resolveRepresentation` compliant with W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
+ * Additionally, this function returns an id-only DID document in the case where a DID has been deleted or upgraded.
+ * If a DID is invalid or has not been registered, this is indicated by the `error` property on the `didResolutionMetadata`.
+ *
+ * @param did The DID to resolve.
+ * @param resolutionOptions The resolution options accepted by the `resolveRepresentation` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
+ * @param resolutionOptions.accept The content type accepted by the requesting client.
+ * @returns The resolution result for the `resolveRepresentation` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-resolution).
+ */
+export async function resolveRepresentation(
+  did: DidUri,
+  { accept }: DereferenceOptions<SupportedContentType> = {
+    accept: DID_JSON,
+  }
+): Promise<RepresentationResolutionResult<SupportedContentType>> {
+  if (!isValidContentType(accept)) {
+    return {
+      didResolutionMetadata: {
+        error: 'representationNotSupported',
+      },
+      didDocumentMetadata: {},
+    }
+  }
+
+  const { didDocumentMetadata, didResolutionMetadata, didDocument } =
+    await resolve(did)
+
+  if (didDocument === undefined) {
+    return {
+      // Metadata is the same, since the `representationNotSupported` is already accounted for above.
+      didResolutionMetadata,
+      didDocumentMetadata,
+    } as RepresentationResolutionResult<SupportedContentType>
+  }
+
+  const bufferInput = (() => {
+    if (accept === 'application/did+json') {
+      return JSON.stringify(didDocument)
+    }
+    if (accept === 'application/did+ld+json') {
+      const jsonLdDoc: JsonLd<DidDocument> = {
+        ...didDocument,
+        '@context': [W3C_DID_CONTEXT_URL, KILT_DID_CONTEXT_URL],
+      }
+      return JSON.stringify(jsonLdDoc)
+    }
+    // contentType === 'application/did+cbor
+    return cbor.encode(didDocument)
+  })()
+
+  return {
+    didDocumentMetadata,
+    didResolutionMetadata,
+    didDocumentStream: Buffer.from(bufferInput),
+  } as RepresentationResolutionResult<SupportedContentType>
+}
+
+type InternalDereferenceResult = {
+  contentStream?: DereferenceContentStream
+  contentMetadata: DereferenceContentMetadata
+}
+
+async function dereferenceInternal(
+  didUrl: DidUri | DidUrl,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  dereferenceOptions: DereferenceOptions<SupportedContentType>
+): Promise<InternalDereferenceResult> {
+  const { did, fragment } = parse(didUrl)
+
+  const { didDocument, didDocumentMetadata } = await resolve(did)
+
+  if (fragment === undefined) {
+    return {
+      contentStream: didDocument,
+      contentMetadata: didDocumentMetadata,
+    }
+  }
+  const dereferencedResource = (() => {
+    const verificationMethod = didDocument?.verificationMethod?.find(
+      (vm) => vm.id === fragment
+    )
+    if (verificationMethod !== undefined) {
+      return verificationMethod
+    }
+
+    const service = didDocument?.service?.find((s) => s.id === fragment)
+    return service
+  })()
+
+  return {
+    contentStream: dereferencedResource,
+    contentMetadata: {},
+  }
+}
+
+/**
+ * Implementation of `dereference` compliant with W3C DID specifications (https://www.w3.org/TR/did-core/#did-url-dereferencing).
+ * If a DID URL is invalid or has not been registered, this is indicated by the `error` property on the `dereferencingMetadata`.
+ *
+ * @param didUrl The DID URL to dereference.
+ * @param resolutionOptions The resolution options accepted by the `dereference` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-url-dereferencing).
+ * @param resolutionOptions.accept The content type accepted by the requesting client.
+ * @returns The resolution result for the `dereference` function as specified in the W3C DID specifications (https://www.w3.org/TR/did-core/#did-url-dereferencing).
+ */
+export async function dereference(
+  didUrl: DidUri | DidUrl,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  { accept }: DereferenceOptions<SupportedContentType> = {
+    accept: DID_JSON,
+  }
+): Promise<DereferenceResult<SupportedContentType>> {
+  // The spec does not include an error for unsupported content types for dereferences
+  const contentType = isValidContentType(accept) ? accept : DID_JSON
+
+  try {
+    validateUri(didUrl)
+  } catch (error) {
+    return {
+      dereferencingMetadata: {
+        error: 'invalidDidUrl',
+      },
+      contentMetadata: {},
+    }
+  }
+
+  const resolutionResult = await dereferenceInternal(didUrl, {
+    accept: contentType,
+  })
+
+  const { contentMetadata, contentStream } = resolutionResult
+
+  if (contentStream === undefined) {
+    return {
+      dereferencingMetadata: {
+        error: 'notFound',
+      },
+      contentMetadata,
+    }
+  }
+
+  const stream = (() => {
+    if (contentType === 'application/did+json') {
+      return contentStream
+    }
+    if (contentType === 'application/did+ld+json') {
+      return {
+        ...contentStream,
+        '@context': [W3C_DID_CONTEXT_URL, KILT_DID_CONTEXT_URL],
+      }
+    }
+    // contentType === 'application/did+cbor'
+    return Buffer.from(cbor.encode(contentStream))
+  })()
+
+  return {
+    dereferencingMetadata: {
+      contentType,
+    },
+    contentMetadata,
+    contentStream: stream,
+  }
+}
+
+/**
+ * Fully-fledged default resolver capable of resolving DIDs in their canonical form, encoded for a specific content type, and of dereferencing parts of a DID Document according to the dereferencing specification.
+ */
+export const resolver: DidResolver<SupportedContentType> = {
+  resolve,
+  resolveRepresentation,
+  dereference,
 }
