+++ packages/did/src/DidResolver/_DidResolver.spec.ts
