+++ packages/types/src/index.ts
@@ -24,10 +24,5 @@ export * from './Credential.js'
 export * from './DidDocument.js'
 export * from './CryptoCallbacks.js'
 export * from './DidResolver.js'
-export * from './DidDocumentExporter.js'
 export * from './PublicCredential.js'
 export * from './Imported.js'
-
-export * as DidDocumentV2 from './DidDocumentV2.js'
-export * as DidResolverV2 from './DidResolverV2.js'
-export * as CryptoCallbacksV2 from './CryptoCallbacksV2.js'
